feat(stories): support relative links
