+++ src/scripts/components/stories/story-content/story-content.styl
@@ -26,6 +26,11 @@
     font-size: emCalc(11px)
     font-family: NotesEsa
     line-height: normal
+  
+  a
+    margin: 1em 0
+    color: $textColor
+    text-decoration: none
 
 .shortTextContent
   overflow-y: hidden
